+++ Editor/Editors/CinemachineHardLookAtEditor.cs
@@ -31,13 +31,9 @@ namespace Unity.Cinemachine.Editor
             if (Target == null || !CinemachineCorePrefs.ShowInGameGuides.Value || !Target.isActiveAndEnabled)
                 return;
 
-            // Don't draw the guides if rendering to texture
-            if (brain == null || brain.OutputCamera == null
-                    || (brain.OutputCamera.activeTexture != null && CinemachineBrain.ActiveBrainCount > 1))
-                return;
-
             var vcam = Target.VirtualCamera;
-            if (!brain.IsValidChannel(vcam))
+            if (brain == null || brain != CinemachineCore.FindPotentialTargetBrain(vcam)
+                || (brain.OutputCamera.activeTexture != null && CinemachineBrain.ActiveBrainCount > 1))
                 return;
 
             bool isLive = targets.Length <= 1 && brain.IsLiveChild(vcam, true);