+++ Editor/Obsolete/CinemachineComposerEditor.cs
@@ -56,12 +56,9 @@ namespace Unity.Cinemachine.Editor
             if (!VcamStageEditor.ActiveEditorRegistry.IsActiveEditor(this))
                 return;
 
-            // Don't draw the guides if rendering to texture
-            if (brain == null || (brain.OutputCamera.activeTexture != null && CinemachineBrain.ActiveBrainCount > 1))
-                return;
-
             var vcam = Target.VirtualCamera;
-            if (!brain.IsLiveChild(vcam))
+            if (brain == null || brain != CinemachineCore.FindPotentialTargetBrain(vcam)
+                || (brain.OutputCamera.activeTexture != null && CinemachineBrain.ActiveBrainCount > 1))
                 return;
 
             // Screen guides