+++ Editor/Obsolete/CinemachineFramingTransposerEditor.cs
@@ -124,11 +124,9 @@ namespace Unity.Cinemachine.Editor
             if (!VcamStageEditor.ActiveEditorRegistry.IsActiveEditor(this))
                 return;
 
-            if (brain == null || (brain.OutputCamera.activeTexture != null && CinemachineBrain.ActiveBrainCount > 1))
-                return;
-
             var vcam = Target.VirtualCamera;
-            if (!brain.IsValidChannel(vcam))
+            if (brain == null || brain != CinemachineCore.FindPotentialTargetBrain(vcam)
+                || (brain.OutputCamera.activeTexture != null && CinemachineBrain.ActiveBrainCount > 1))
                 return;
 
             // Screen guides