+++ Runtime/Core/CinemachinePropertyAttribute.cs
@@ -99,7 +99,7 @@ namespace Unity.Cinemachine
     /// <summary>
     /// Used for custom drawing in the inspector.  Inspector will show a foldout with the asset contents
     /// </summary>
-    /// GML TODO: delete this attribute
+    // GML TODO: delete this attribute
     public sealed class CinemachineEmbeddedAssetPropertyAttribute : PropertyAttribute 
     {
         /// <summary>If true, inspector will display a warning if the embedded asset is null</summary>
@@ -163,7 +163,7 @@ namespace Unity.Cinemachine
     }
 
     /// <summary>
-    /// Attribute applied to a CinemachineVirtualCameraBase property to produce
+    /// Attribute applied to a CinemachineCameraManagerBase property to produce
     /// a child camera selector in the inspector.
     /// </summary>
     public sealed class ChildCameraPropertyAttribute : PropertyAttribute {}